pyproject.toml
[tool.poetry]
name = "carla_experiments"
version = "0.1.0"
description = "Testing repo for the carla api"
authors = ["Ulrik Røsby <ulrik@rosby.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.10"
carla = "0.9.15"
numpy = "1.24"
tqdm = "^4.66.1"
mypy = "^1.6.0"
command = "^0.1.0"
click = "^8.1.7"
wandb = "^0.16.0"
torchmetrics = "^1.2.1"
matplotlib = "3.6"
pandas = "1.4.3"
opencv-python = "^4.9.0.80"
ffmpeg-python = "^0.2.0"
pymap3d = "^3.0.1"
scipy = "1.10"
pyproj = "3.3.0"
efficientnet-pytorch = "^0.7.1"
flake8 = "^7.0.0"
black = "^24.3.0"
pycapnp = "^1.3.0"
strenum = "^0.4.15"
onnx2pytorch = "^0.4.1"
onnx2torch = "^1.5.14"
onnx-pytorch = "^0.1.5"
onnxruntime-gpu = "^1.17.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
